pref: reduce rerender in candidate node
